Complete Assignment 1: Implemented all required functionalities and optimizations.

+++ Assignments/Assignment1/complex.h
@@ -12,8 +12,8 @@
 
 struct Complex
 {
-    float real;
-    float imag;
+    double real;
+    double imag;
 };
 
 // Function prototypes
